% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_motif.R
\name{add_MotifEnrichmentResults}
\alias{add_MotifEnrichmentResults}
\title{combine two MotifEnrichmentResults dervied from unique region sets and update statistics}
\usage{
add_MotifEnrichmentResults(me_res1, me_res2, pwmobj)
}
\arguments{
\item{me_res1}{MotifEnrichmentResults}

\item{me_res2}{MotifEnrichmentResults}
}
\value{
MotifEnrichmentResults of combined inputs
}
\description{
combine two MotifEnrichmentResults dervied from unique region sets and update statistics
}
