% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cache.R
\name{bfcif}
\alias{bfcif}
\title{Conditionally run and cache FUN if results aren't already cached.  Returns result of FUN.}
\usage{
bfcif(bfc, rname, FUN, force_overwrite = FALSE)
}
\arguments{
\item{bfc}{BiocFileCache}

\item{force_overwrite}{}
}
\description{
Conditionally run and cache FUN if results aren't already cached.  Returns result of FUN.
}
