% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_crossCorrRle.R
\name{calcCorrMetrics}
\alias{calcCorrMetrics}
\title{Title}
\usage{
calcCorrMetrics(bam_file, qgr, frag_min, frag_max, bam_md5 = NULL,
  qgr_md5 = NULL, cache_path = "~/.cache_peakrefine",
  cach_version = "v1", force_overwrite = FALSE,
  n_splits = getOption("mc.cores", 1L))
}
\arguments{
\item{n_splits}{}
}
\description{
Title
}
