% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_motif.R
\name{calcMotifEnrichment}
\alias{calcMotifEnrichment}
\title{Title}
\usage{
calcMotifEnrichment(corr_res, base_gr, qdt, todo_groups, pwm, seq, bam_md5,
  nbases = 200, ncores = 8, cache_path = "~/.cache_peakrefine",
  force_overwrite = FALSE, force_overwrite_pre = FALSE,
  cach_version = "v1")
}
\arguments{
\item{cach_version}{}
}
\description{
Title
}
