% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_crossCorrRle.R
\name{calcSCCMetrics}
\alias{calcSCCMetrics}
\title{Title}
\usage{
calcSCCMetrics(bam_file, qgr, frag_sizes, fetch_size = 3 *
  max(frag_sizes), bam_md5 = NULL, qgr_md5 = NULL,
  cache_path = "~/.cache_peakrefine", cach_version = "v1",
  force_overwrite = FALSE, n_splits = getOption("mc.cores", 1L))
}
\arguments{
\item{n_splits}{}
}
\description{
Title
}
