% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_crossCorrExtension.R
\name{crossCorrByExtension}
\alias{crossCorrByExtension}
\title{Title}
\usage{
crossCorrByExtension(bam_file, qgr, nbest = 20, qual_metric = "qValue",
  max_dupes = 1, frag_min = 50, frag_max = 250, step = 10,
  small_step = 1, include_plots = TRUE)
}
\arguments{
\item{include_plots}{}
}
\description{
Title
}
