% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_crossCorrRle.R
\name{crossCorrByRle}
\alias{crossCorrByRle}
\title{Calculate cross correlation by extending reads}
\usage{
crossCorrByRle(bam_file, query_gr, max_dupes = 1,
  fragment_sizes = 50:300, read_length = NULL, include_plots = TRUE)
}
\arguments{
\item{bam_file}{character. Path to .bam file, must have index at .bam.bai.}

\item{query_gr}{GRanges.  Regions to calculate cross correlation for.}

\item{max_dupes}{integer.  Duplicate reads above this value will be removed.}

\item{fragment_sizes}{integer.  fragment size range to search for maximum
correlation.}

\item{read_length}{integer. Any values outside fragment_range that must be
searched.  If not supplied will be determined from bam_file.  Set as NA
to disable this behavior.}

\item{include_plots}{logical. Should plots be included in output?}
}
\value{
named list of results
}
\description{
Calculate cross correlation by extending reads
}
\examples{
bam_file = system.file("extdata", "MCF10A_CTCF.random5.bam", package = "peakrefine")
np = system.file("extdata", "MCF10A_CTCF.random5.narrowPeak", package = "peakrefine")
query_gr = rtracklayer::import(np, format = "narrowPeak")
crossCorrByExtension(bam_file, query_gr[1:2], fragment_range = 50:250,
step = 50, small_step = 10)
}
