% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_crossCorrShift.R
\name{crossCorrByShift}
\alias{crossCorrByShift}
\title{Title}
\usage{
crossCorrByShift(bam_file, qgr, nbest = 20, revbest = FALSE,
  qual_metric = "qValue", max_dupes = 1, shift_min = 0,
  shift_max = 250, step = 10, small_step = 1, include_plots = TRUE)
}
\arguments{
\item{include_plots}{}
}
\description{
Title
}
