% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{.calc_cross_corr}
\alias{.calc_cross_corr}
\title{Calculate correlation between signal on DNA strands}
\usage{
.calc_cross_corr(reads_dt, query_gr, frag_len = NA, window_size = 1)
}
\arguments{
\item{reads_dt}{data.table containing read spans. mandatory columns are
which_label, strand, start, end}

\item{query_gr}{GRanges of regions of interest}

\item{frag_len}{fragment size to extend reads to.  frag_len of NA causes read
extension to be skipped, ie. raw read pileup.}

\item{window_size}{integer, resolution used to represent coverage.  window_size of 10
retrieves coverage every 10 bp.}
}
\value{
data.table of correlation per region in query_gr
}
\description{
Calculate correlation between signal on DNA strands
}
\examples{
bam_file = system.file("extdata", "MCF10A_CTCF.random5.bam", package = "peakrefine")
np = system.file("extdata", "MCF10A_CTCF.random5.narrowPeak", package = "peakrefine")
qgr = rtracklayer::import(np, format = "narrowPeak")
bam_dt = peakrefine:::.fetch_bam_stranded(bam_file, qgr)
peakrefine:::.calc_cross_corr(bam_dt, qgr,  frag_len = NA) #read correlation
peakrefine:::.calc_cross_corr(bam_dt, qgr, frag_len = 150) #fragment 150 bp correlation
}
