% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{.calc_stranded_coverage}
\alias{.calc_stranded_coverage}
\title{Calculate coverage per strand}
\usage{
.calc_stranded_coverage(reads_dt, query_gr, window_size = 1)
}
\arguments{
\item{reads_dt}{data.table containing read spans. mandatory columns are
which_label, strand, start, end}

\item{query_gr}{GRanges of regions of interest}

\item{window_size}{resolution used to represent coverage.  window_size of 10
retrieves coverage every 10 bp.}
}
\value{
data.table containing stranded read coverage
}
\description{
Calculate coverage per strand
}
\examples{
bam_file = system.file("extdata", "MCF10A_CTCF.random5.bam", package = "peakrefine")
np = system.file("extdata", "MCF10A_CTCF.random5.narrowPeak", package = "peakrefine")
qgr = rtracklayer::import(np, format = "narrowPeak")
bam_dt = .fetch_bam_stranded(bam_file, qgr)
cov_dt = .calc_stranded_coverage(bam_dt, qgr)


}
