% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{.extend_reads}
\alias{.extend_reads}
\title{extend reads to specified length in a strand sensitve way}
\usage{
.extend_reads(reads_dt, frag_len)
}
\arguments{
\item{reads_dt}{data.table. Contains at least 3 mandatory columns: strand, start,
and end.}

\item{frag_len}{integer. the fragment length to extend to}
}
\value{
data.table after extending to fragment length
}
\description{
extend reads to specified length in a strand sensitve way
}
\examples{
bam_file = system.file("extdata",
  "MCF10A_CTCF.random5.bam", package = "peakrefine")
bam_input = system.file("extdata",
  "MCF10A_input.random5.bam", package = "peakrefine")
np = system.file("extdata",
  "MCF10A_CTCF.random5.narrowPeak", package = "peakrefine")
query_gr = rtracklayer::import(np, format = "narrowPeak")
reads_dt = peakrefine:::.fetch_bam_stranded(bam_file, query_gr)
peakrefine:::.extend_reads(reads_dt, 150)
}
