% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{.fetch_bam_stranded}
\alias{.fetch_bam_stranded}
\title{.fetch_bam_stranded}
\usage{
.fetch_bam_stranded(bam_f, qgr, max_dupes = Inf, ...)
}
\arguments{
\item{bam_f}{character. a .bam file, must have index file in same directory:
.bam.bai}

\item{qgr}{GRanges, regions to retrieve reads for.}

\item{max_dupes}{numeric, maximum positional duplicates allowed.  See
\code{\link{.remove_duplicates}}}

\item{...}{additional arguments passed to
\code{\link[Rsamtools]{ScanBamParam}}}
}
\value{
a tidy data.table of read pileup coverage from bam_f for regions in
  qgr
}
\description{
.fetch_bam_stranded
}
\examples{
bam_file = system.file("extdata",
  "MCF10A_CTCF.random5.bam", package = "peakrefine")
bam_input = system.file("extdata",
  "MCF10A_input.random5.bam", package = "peakrefine")
np = system.file("extdata",
  "MCF10A_CTCF.random5.narrowPeak", package = "peakrefine")
qgr = rtracklayer::import(np, format = "narrowPeak")
.fetch_bam_stranded(bam_file, qgr)

}
