% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{.remove_duplicates}
\alias{.remove_duplicates}
\title{remove positional read duplicates from data.table}
\usage{
.remove_duplicates(reads_dt, max_dupes)
}
\arguments{
\item{reads_dt}{a data.table containing read span information.  required
columns are which_label, start, end, and strand.}

\item{max_dupes}{integer >= 1, the maximum allowed duplicates per position.}
}
\value{
data.table with reads in excess of max_dupes removed
}
\description{
remove positional read duplicates from data.table
}
\examples{
library(data.table)
n = 4
dt = data.table(
  which_label = seq_len(n),
  seqnames = "chr1",
  strand = c("+", "-"),
  start = seq_len(n),
  end = seq_len(n) + 10
)
dt[, width := end - start + 1]
make_dupes = seq_len(nrow(dt))
make_dupes = rep(make_dupes, make_dupes)
dt = dt[make_dupes]
lapply(1: n, function(x)peakrefine:::.remove_duplicates(dt, max_dupes = x))
}
