% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{.view_gr_by_window_sample}
\alias{.view_gr_by_window_sample}
\title{view GRanges with score at specified window size by sampling values}
\usage{
.view_gr_by_window_sample(score_gr, query_gr, window_size,
  anchor = c("center", "center_unstranded", "left",
  "left_unstranded")[1])
}
\arguments{
\item{score_gr}{GRanges with score mcol}

\item{query_gr}{GRanges that was used to fecth score_gr}

\item{window_size}{ineger. desired window size.}

\item{anchor}{character, anchoring method. one of "center", "center_unstranded",
"left", "left_unstranded"}
}
\value{
data.table of scores stored in y value sample every window_size
}
\description{
view GRanges with score at specified window size by sampling values
}
\examples{
library(GenomicRanges)
bam_file = system.file("extdata", "MCF10A_CTCF.random5.bam", package = "peakrefine")
bam_input = system.file("extdata", "MCF10A_input.random5.bam", package = "peakrefine")
np = system.file("extdata", "MCF10A_CTCF.random5.narrowPeak", package = "peakrefine")
qgr = rtracklayer::import(np, format = "narrowPeak")
strand(qgr) = c("+", "-", "-", "+", "-")

reads_dt = peakrefine:::.fetch_bam_stranded(bam_file, qgr)
ext_cov = coverage(split(
    GRanges(reads_dt[strand == "+"]), reads_dt$which_label))
score_gr = GRanges(ext_cov)
peakrefine:::.view_gr_by_window_sample(score_gr, qgr, 1,
                                 anchor = "center_unstranded")
}
