% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_crossCorrRle.R
\name{getReadLength}
\alias{getReadLength}
\title{get the read length from a bam file}
\usage{
getReadLength(bam_file, query_gr)
}
\arguments{
\item{bam_file}{character. Path to .bam file, must have index at .bam.bai.}

\item{query_gr}{GRanges.  Regions to calculate cross correlation for.  Only first 10 actually used.}
}
\value{
numeric.  the most common read length observed in bam.
}
\description{
get the read length from a bam file
}
\examples{
bam_file = system.file("extdata", "MCF10A_CTCF.random5.bam", package = "peakrefine")
np = system.file("extdata", "MCF10A_CTCF.random5.narrowPeak", package = "peakrefine")
query_gr = rtracklayer::import(np, format = "narrowPeak")
getReadLength(bam_file, query_gr)

}
