% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utils.R
\name{harmonize_seqlengths}
\alias{harmonize_seqlengths}
\title{Ensure that regions in GRanges are compatible wtih seqlengths of bam_file}
\usage{
harmonize_seqlengths(gr, bam_file)
}
\arguments{
\item{gr}{GRanges that possibly extend to invalid start or end values.}

\item{bam_file}{bam_file to retrieve seqlengths info from}
}
\value{
GRanges shifted such that all start at 1 or greater and end prior to chromosome end
}
\description{
Ensure that regions in GRanges are compatible wtih seqlengths of bam_file
}
