% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_peakrefine.R
\docType{class}
\name{PeakRefiner-class}
\alias{PeakRefiner-class}
\title{class to handle peak refinement}
\description{
class to handle peak refinement
}
\section{Slots}{

\describe{
\item{\code{peak_set}}{GRanges.}

\item{\code{bam_treat_file}}{character.}

\item{\code{bam_input_file}}{character.}

\item{\code{fragment_lengths}}{integer.}

\item{\code{fl2color}}{character.}

\item{\code{pwm}}{PWMLognBackground.}

\item{\code{target_pwm_names}}{character.}

\item{\code{output_prefix}}{character.}

\item{\code{plots}}{list.}
}}

