% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_peakrefine.R
\name{PeakRefiner}
\alias{PeakRefiner}
\title{HiC_matrix constructor}
\usage{
PeakRefiner(peak_set, bam_treat_file, bam_input_file, pwm,
  target_pwm_names, fragment_lengths = NULL, color_overrides = NULL,
  auto_frag_len_FUN = NULL, output_prefix = NULL)
}
\arguments{
\item{matrix_file}{HiC-Pro matrix file}

\item{regions_file}{HiC-Pro region bed file matching matrix file}

\item{parameters}{HiC_parameters object}
}
\description{
HiC_matrix constructor
}
