% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_motif.R
\name{pre_motif}
\alias{pre_motif}
\title{Calculate a large motif enrichment once for later rounds of
subsetting and analysis}
\usage{
pre_motif(qgr, pwm = PWMLogn.hg19.MotifDb.Hsap, nbases = 200,
  ncores = 32)
}
\arguments{
\item{ncores}{}
}
\description{
Calculate a large motif enrichment once for later rounds of
subsetting and analysis
}
