% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_motif.R
\name{PWMEnrich.affinitySequenceSet}
\alias{PWMEnrich.affinitySequenceSet}
\title{a non-exported method from PWMEnrich, here for stability}
\usage{
PWMEnrich.affinitySequenceSet(scores, seq.len, pwm.len)
}
\description{
a non-exported method from PWMEnrich, here for stability
}
