% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_motif.R
\name{PWMEnrich.cloverScore}
\alias{PWMEnrich.cloverScore}
\title{a non-exported method from PWMEnrich, here for stability}
\usage{
PWMEnrich.cloverScore(scores, lr3 = FALSE, verbose = FALSE)
}
\description{
a non-exported method from PWMEnrich, here for stability
}
