% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{ssvCrossCorr}
\alias{ssvCrossCorr}
\title{Title}
\usage{
ssvCrossCorr(bam_file, qgr, nbest = 20, revbest = FALSE,
  qual_metric = "qValue", max_dupes = 1, shift_min = 0,
  shift_max = 250, step = 10, include_plots = TRUE)
}
\arguments{
\item{include_plots}{}
}
\description{
Title
}
