% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{ssvStrandCorr}
\alias{ssvStrandCorr}
\title{Title}
\usage{
ssvStrandCorr(bam_file, qgr, nbest = 20, qual_metric = "qValue",
  max_dupes = 1, frag_min = 50, frag_max = 250, step = 10,
  include_plots = TRUE)
}
\arguments{
\item{include_plots}{}
}
\description{
Title
}
