% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{ssvStrandCorrFull}
\alias{ssvStrandCorrFull}
\title{Title}
\usage{
ssvStrandCorrFull(bam_file, qgr, fragLen, qual_metric = "id",
  max_dupes = 1, ncores = 1, output_withGRanges = TRUE)
}
\arguments{
\item{ncores}{}
}
\description{
Title
}
