% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PWM_modify.R
\name{subset_MotifEnrichmentResults}
\alias{subset_MotifEnrichmentResults}
\title{subset MotifEnrichmentResults by region and update statistics}
\usage{
subset_MotifEnrichmentResults(me_res, k, pwmobj)
}
\arguments{
\item{me_res}{MotifEnrichmentResults}

\item{k}{logical or integer (as from which(logical)) to subset results by region.}
}
\value{
MotifEnrichmentResults subsetted by k
}
\description{
subset MotifEnrichmentResults by region and update statistics
}
