% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_calc_fragLen.R
\name{viewGRangesWinSample_dt}
\alias{viewGRangesWinSample_dt}
\title{Title}
\usage{
viewGRangesWinSample_dt(score_gr, qgr, window_size, anchor = c("center",
  "center_unstranded", "left", "left_unstranded")[1])
}
\arguments{
\item{anchor}{}
}
\description{
Title
}
