% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_motif.R
\name{calcMotifEnrichment}
\alias{calcMotifEnrichment}
\title{Title}
\usage{
calcMotifEnrichment(corr_res, base_gr, qdt, pwm, seq, bam_md5,
  nbases = 200, cache_path = "~/.cache_peakrefine",
  cach_version = "v1")
}
\arguments{
\item{cach_version}{}
}
\description{
Title
}
