% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_crossCorrExtension.R
\name{crossCorrByExtensionFull}
\alias{crossCorrByExtensionFull}
\title{Title}
\usage{
crossCorrByExtensionFull(bam_file, qgr, fragLen, qual_metric = "id",
  max_dupes = 1, ncores = 1, output_withGRanges = TRUE)
}
\arguments{
\item{ncores}{}
}
\description{
Title
}
