% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_peakrefine.R
\docType{methods}
\name{initialize,PeakRefiner-method}
\alias{initialize,PeakRefiner-method}
\title{initialize a new PeakRefiner}
\usage{
\S4method{initialize}{PeakRefiner}(.Object, peak_set, bam_treat_file,
  bam_input_file, pwm, target_pwm_names, fragment_lengths = NULL,
  color_overrides = NULL, color_default = "black",
  auto_frag_len_FUN = NULL, output_prefix = NULL)
}
\arguments{
\item{PeakRefiner}{empty PeakRefiner}
}
\value{
valid PeakRefiner
}
\description{
initialize a new PeakRefiner
}
