% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PWM_modify.R
\name{pre_motif}
\alias{pre_motif}
\title{Calculate a large motif enrichment once for later rounds of
subsetting and analysis}
\usage{
pre_motif(query_gr, pwm, seq_reference, nbases = 200, ncores = 32,
  verbose = TRUE)
}
\arguments{
\item{query_gr}{GRanges.  Regions to calculate cross correlation for.}

\item{pwm}{a position weight matrix object to pass to
\code{\link[PWMEnrich]{motifEnrichment}}.}

\item{nbases}{number of bases around center of query_gr regions to consider
for motif enrichment.}

\item{ncores}{number of computer cores to use.}
}
\value{
MotifEnrichmentResults for each region in \code{query_gr} and every
  motif in \code{pwm}
}
\description{
Calculate a large motif enrichment once for later rounds of
subsetting and analysis
}
