% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_motif.R
\name{score_motif}
\alias{score_motif}
\title{Assembles several metrics for a set of peaks from:
1) the original peak call
2) the ChIP-seq bam file
3) an optional input (no pull-down) bam file}
\usage{
score_motif(bam_file, bam_input, qgr, fl, motif_res = NULL,
  pwm = PWMLogn.hg19.MotifDb.Hsap, include_fl_independent = FALSE,
  out_dir = file.path("~/R/peakrefine/cache", sub(".bam", "",
  basename(bam_file))), nbases = 200, ncores = 32)
}
\arguments{
\item{ncores}{}
}
\description{
Stratify by each metric and conduct motif enrichment for each.
}
